requirements.txt
    Flask==3.0.0
gunicorn==21.2.0sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0